+++ mbo/container/any_scan.h
@@ -40,7 +40,8 @@ namespace mbo::container {
 // HOWEVER: This types is slower as the aformentioned wrappers, as it need more abstraction layers
 // to accomplish the independence of the container type.
 //
-// The wrappers of this type can only be instantiated/created through `MakeAnyScan`.
+// The wrappers of this type can only be instantiated/created through `MakeAnyScan` or directly from
+// a `std::initializer_list` argument.
 //
 // The implementqtion is type-safe and does not use any `reinterpret_cast` of other unnecessary
 // tricks.
@@ -75,10 +76,14 @@ namespace mbo::container {
 template<typename ValueType, typename DifferenceType = std::ptrdiff_t>
 class AnyScan;
 
-// A `ConvertingScan` is very similar to `AnyScan`, however it allows scanning of any container whose `value_types` - as
-// returned by its const_iterator - can be copy-converted into the specified type. This in particular allows scanning
-// containers of any string-like type as either `std::string` or `std::string_view`. The emphasis is on COPY as this
-// kind of conversion requires return by value. As such the `ConvertingScan`'s iterators do not have `operator->()`.
+// A `ConvertingScan` is very similar to `AnyScan`, however it allows scanning of any container
+// whose `value_types` - as returned by its const_iterator - can be copy-converted into the
+// specified type. This in particular allows scanning containers of any string-like type as either
+// `std::string` or `std::string_view`. The emphasis is on COPY as this kind of conversion requires
+// return by value. As such the `ConvertingScan`'s iterators do not have `operator->()`.
+//
+// The wrappers of this type can only be instantiated/created through `MakeConvertingScan` or
+// directly from a `std::initializer_list` argument.
 //
 // Example:
 //
@@ -91,8 +96,9 @@ class AnyScan;
 // }
 // ```
 //
-// Here the caller used `std::vector` with `value_type` set to `std::string`, but the target `Report` wants the elements
-// converted to `std::string_view`, so this had to be a `ConvertingScan` that handles the type difference.
+// Here the caller used `std::vector` with `value_type` set to `std::string`, but the target
+// `Report` wants the elements converted to `std::string_view`, so this had to be a `ConvertingScan`
+// that handles the type difference.
 template<typename ValueType, typename DifferenceType = std::ptrdiff_t>
 class ConvertingScan;
 
@@ -102,13 +108,13 @@ template<typename ValueType, typename DifferenceType = std::ptrdiff_t, bool Acce
 class AnyScanImpl;
 
 template<::mbo::types::ContainerHasInputIterator Container, bool AccessByRef, typename AccessType>
-requires (!AccessByRef || std::same_as<AccessType, void>)
+requires(!AccessByRef || std::same_as<AccessType, void>)
 struct AnyScanTypeImpl {
   using RawContainer = std::remove_cvref_t<Container>;
   using ValueType = std::conditional_t<AccessByRef, typename RawContainer::value_type, AccessType>;
   using ActualDifferenceType = ::mbo::types::GetDifferenceType<RawContainer>;
-  using DifferenceType = std::conditional_t<
-      std::convertible_to<ActualDifferenceType, std::ptrdiff_t>, std::ptrdiff_t, ActualDifferenceType>;
+  using DifferenceType = std::
+      conditional_t<std::convertible_to<ActualDifferenceType, std::ptrdiff_t>, std::ptrdiff_t, ActualDifferenceType>;
 
   using type = std::conditional_t<
       AccessByRef,
@@ -175,9 +181,9 @@ class AnyScanImpl {
 
   template<::mbo::types::ContainerHasInputIterator Container>
   requires(
-    !AccessByRef &&
-    std::constructible_from<AccessType, ::mbo::types::ContainerConstIteratorValueType<Container>> &&
-    std::constructible_from<value_type, AccessType>)
+      !AccessByRef  // This is the ConvertingScan constructor
+      && std::constructible_from<AccessType, ::mbo::types::ContainerConstIteratorValueType<Container>>
+      && std::constructible_from<value_type, AccessType>)
   explicit AnyScanImpl(Container&& container)  // NOLINT(bugprone-forwarding-reference-overload)
       : clone_([container = std::forward<Container>(container)] {
           using RawContainer = std::remove_cvref_t<Container>;
@@ -206,7 +212,7 @@ class AnyScanImpl {
     const_reference operator*() const noexcept
     requires AccessByRef
     {
-      // We check both the presence of `func_` as well as calling the actual `more` function.
+      // We check both the presence of `funcs_` as well as calling the actual `more` function.
       // That means we bypass any protection an iterator may have, but we can make this function
       // `noexcept` assuming the iterator is noexcept for access. On the other hand we expect that
       // out of bounds access may actually raise. So we effectively side step such exceptions.
@@ -280,10 +286,15 @@ class AnyScan : public container_internal::AnyScanImpl<ValueType, DifferenceType
   using AnyScanImpl::AnyScanImpl;
 
  public:
-  // The constructor is private and the static constructor must be used.
+  // Only the initializer_list constructor is public.
+  // The other constructors are private and the static constructor must be used.
   // HOWEVER, prefer the free standing function `MakeAnyScan`.
   template<::mbo::types::ContainerHasInputIterator Container>
-  static AnyScan InternalMakeAnyScan(Container&& container) noexcept { return AnyScan(std::forward<Container>(container)); }
+  static AnyScan InternalMakeAnyScan(Container&& container) noexcept {
+    return AnyScan(std::forward<Container>(container));
+  }
+
+  AnyScan(const std::initializer_list<ValueType>& data) : AnyScanImpl(data) {}
 
   using AnyScanImpl::begin;
   using AnyScanImpl::end;
@@ -305,10 +316,15 @@ class ConvertingScan : public container_internal::AnyScanImpl<ValueType, Differe
   using AnyScanImpl::AnyScanImpl;
 
  public:
-  // The constructor is private and the static constructor must be used.
+  // Only the initializer_list constructor is public.
+  // The other constructors are private and the static constructor must be used.
   // HOWEVER, prefer the free standing function `MakeConvertingScan`.
   template<::mbo::types::ContainerHasInputIterator Container>
-  static ConvertingScan InternalMakeAnyScan(Container&& container) noexcept { return ConvertingScan(std::forward<Container>(container)); }
+  static ConvertingScan InternalMakeAnyScan(Container&& container) noexcept {
+    return ConvertingScan(std::forward<Container>(container));
+  }
+
+  ConvertingScan(const std::initializer_list<ValueType>& data) noexcept : AnyScanImpl(data) {}
 
   using AnyScanImpl::begin;
   using AnyScanImpl::end;